refactor: extract HelperPanelControl to shared components, use for experiment detail
Move InsightsHelperPanelControl visual logic to @inngest/components/HelperPanelControl.
Replace custom SidebarRail in experiment detail with the generic component.
Fix layout: icon rail now on far right, sidebar panel slides in from its left.
