+++ ui/packages/components/src/HelperPanelControl/index.ts
@@ -1,0 +1,1 @@
+export { HelperPanelControl, type HelperItem } from './HelperPanelControl';
